+++ module/documents/active-effect/torgActiveEffect.js
@@ -152,7 +152,7 @@ export default class TorgActiveEffect extends foundry.documents.ActiveEffect {
     return !this.disabled && this.system.transferOnOutcome && this.system.transferTo === 'actor';
   }
 
-  get isTransferrable() {
+  get isTransferable() {
     return !this.disabled && !!this.system.transferOnOutcome;
   }